+++ scripts/git-subtree/custom-fonts/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/custom-fonts `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-custom-fonts.git `
+    dev `
+    --squash