+++ scripts/git-subtree/custom-fonts/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/custom-fonts `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-custom-fonts.git `
+    dev