+++ scripts/git-subtree/custom-fonts/split.ps1
@@ -1,0 +1,3 @@
+git subtree split `
+    -P workspace/plugins/custom-fonts `
+    --rejoin